Rescale fitting pars using logits
When fitting, we used L-BFGS-B in optim() to bound the
parameters to sensible values. We now use
a (inv)logit-based rescaling so that we can do the
optimization with nlm that does not accept bounds. nlm()
seems much more robust to the production of Inf/NaN/etc.
by the likelihood function.